chore: rename `source/ratelimit-header-parser.ts` -> `source/parser.ts`
